restructure debug UI
