--- assets/ui/debug.ui
@@ -1,152 +1,0 @@
-{
-  "type": "DebugHud",
-  "contents": {
-    "type": "relativeLayout",
-    "layoutInfo": {
-      "use-content-width": true,
-      "use-content-height": true,
-      "position-horizontal-center": {},
-      "position-top": {
-        "target": "TOP",
-        "offset": 80
-      }
-    },
-    "contents": [
-      {
-        "type": "rowLayout",
-        "layoutInfo": {
-          "use-content-width": true,
-          "use-content-height": true
-        },
-        "contents": [
-          {
-            "type" : "ColumnLayout",
-            "id": "successColumn",
-            "columns" : 1,
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "Success Times",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "successTimes"
-              }
-            ]
-          },
-          {
-            "type" : "ColumnLayout",
-            "id": "failureColumn",
-            "columns" : 1,
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "Failure Times",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "failureTimes"
-              }
-            ]
-          },
-          {
-            "type" : "ColumnLayout",
-            "columns" : 1,
-            "id": "costsColumn",
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "Costs",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "costs"
-              }
-            ]
-          },
-          {
-            "type" : "ColumnLayout",
-            "columns" : 1,
-            "id": "sizesColumn",
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "Sizes",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "sizes"
-              }
-            ]
-          },
-          {
-            "type" : "ColumnLayout",
-            "columns" : 1,
-            "id": "depthsColumn",
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "depths",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "depths"
-              }
-            ]
-          },
-          {
-            "type" : "ColumnLayout",
-            "columns" : 1,
-            "id": "exploredColumn",
-            "verticalSpacing" : 8,
-            "layoutInfo" : {
-              "use-content-width" : true,
-              "use-content-height" : true
-            },
-            "contents" : [
-              {
-                "type" : "UILabel",
-                "text" : "explored",
-                "family" : "title"
-              },
-              {
-                "type": "UIHistogram",
-                "id": "explored"
-              }
-            ]
-          }
-        ]
-      }
-    ]
-  }
-}
