+++ src/main/java/org/terasology/flexiblepathfinding/debug/ui/DebugScreen.java
@@ -1,0 +1,17 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.flexiblepathfinding.debug.ui;
+
+import org.terasology.flexiblepathfinding.debug.DebugClientSystem;
+import org.terasology.registry.In;
+import org.terasology.rendering.nui.CoreScreenLayer;
+import org.terasology.rendering.nui.databinding.DefaultBinding;
+
+import java.util.Map;
+
+public class DebugScreen extends CoreScreenLayer {
+    @Override
+    public void initialise() {
+
+    }
+}
